requirements.txt
azure-identity>=1.15.0
requests>=2.31.0
PyYAML>=6.0.1
pytest>=7.4.0
